requirements.txt
opencv-python
flask
python-ffmpeg
pyaudio
pyttsx3
elevenlabslib
requests
dropbox
ffpyplayer
soundfile
sounddevice